requirements.txt
gymnasium>=0.29.1
stable-baselines3>=2.0.0
torch>=2.0.0
numpy>=1.21.0
networkx>=3.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.5.0
tensorboard>=2.13.0